Split state mutations into focused handlers
