+++ src/analysis/state/apply_misc.rs
@@ -1,0 +1,55 @@
+use super::{AnalysisState, MutationResult};
+use crate::analysis::mutations::{
+    AlterDatabaseMutation, AlterDomainMutation, CreateDatabaseMutation, DropDatabaseMutation,
+    RefreshMaterializedViewMutation,
+};
+use crate::ast::identifiers::ObjectId;
+
+impl AnalysisState {
+    pub(super) fn apply_refresh_materialized_view(
+        &mut self,
+        _refresh: &RefreshMaterializedViewMutation,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_alter_domain(
+        &mut self,
+        _alter_domain: &AlterDomainMutation,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_check_timeouts(&mut self) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_create_database(
+        &mut self,
+        _create_database: &CreateDatabaseMutation,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_alter_database(
+        &mut self,
+        _alter_database: &AlterDatabaseMutation,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_drop_database(
+        &mut self,
+        _drop_database: &DropDatabaseMutation,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+
+    pub(super) fn apply_vacuum(
+        &mut self,
+        _table_id: &Option<ObjectId>,
+        _is_full: bool,
+    ) -> MutationResult {
+        MutationResult::Applied
+    }
+}
